Cleanup code after adding 3 speed shooter selection.  Go back to 2 speed for driver control, but allow 3 speed for auto.

+++ src/org/wildstang/yearly/auto/programs/LowGoalTest.java
@@ -18,7 +18,7 @@ import org.wildstang.yearly.auto.steps.shooter.StepSetShooterPosition;
 
 public class LowGoalTest extends AutoProgram
 {
-   private double speed;
+   private int speed;
 
    @Override
    protected void defineSteps()
